Moved StopArrivalListItem to Sections/StopArrival/StopArrivalCell
Hide cell separator on non-“list” items
Remove unused code
