+++ OBAKit/Stops/Sections/Generic/MessageButtonItem.swift
@@ -124,6 +124,10 @@ final class MessageButtonCell: OBAListViewCell {
     var showActivityIndicatorOnSelect: Bool = false
     var onTapAction: VoidBlock?
 
+    override var showsSeparator: Bool {
+        return false
+    }
+
     override init(frame: CGRect) {
         super.init(frame: frame)
